refactor(reminders): extract schema helper logic from controller
